MONKEY implement transition between 2 homes types views
