+++ src/store/types.ts
@@ -1,4 +1,7 @@
+import { FetchedVideosState } from "./fetchedVideosState/types";
+
 export interface RootState {
     version: string;
+    fetchedVideosState?: FetchedVideosState;
 }
